+++ FloppyControlApp.net7/WindowsFormsApplication2/WindowsFormsApplication2/MyClasses/Processing/ProcessingTypes/ProcTypeAdaptive2.cs
@@ -8,12 +8,12 @@ namespace FloppyControlApp.MyClasses.Processing.ProcessingTypes
 {
     public partial class ProcessingTypes
     {
-        public byte[] ProcTypeAdaptive2(ProcTypeArgs ProctypeArgs, int ThreadId, ref int Stop)
+        public byte[] ProcTypeAdaptive2(ProcTypeArgs ProctypeArgs, long ThreadId, ref int Stop)
         {
-            int i;
+            long i;
             int value;
 
-            int MINUS, FOURUS, SIXUS, EIGHTUS, start, end;
+            long MINUS, FOURUS, SIXUS, EIGHTUS, start, end;
             float RateOfChange;
             System.Diagnostics.Stopwatch SW = new System.Diagnostics.Stopwatch();
