+++ FloppyControlApp.net7/WindowsFormsApplication2/WindowsFormsApplication2/MyClasses/Processing/ProcessingTypes/ProcTypeArgs.cs
@@ -8,14 +8,14 @@ namespace FloppyControlApp.MyClasses.Processing.ProcessingTypes
 {
     public class ProcTypeArgs
     {
-        public int MINUS { get; set; }
-        public int FOURUS { get; set; }
-        public int SIXUS { get; set; }
-        public int EIGHTUS { get; set; }
+        public long MINUS { get; set; }
+        public long FOURUS { get; set; }
+        public long SIXUS { get; set; }
+        public long EIGHTUS { get; set; }
         public byte[] Rxbuf { get; set; }
         public float[] entropy { get; set; }
-        public int[] Mfmlengths { get; set; }
-        public int[] Progresses { get; set; }
+        public long[] Mfmlengths { get; set; }
+        public long[] Progresses { get; set; }
         public float RateOfChange { get; set; }
 
         public ProcSettings Procsettings { get; set; }
